Los valores se pasan del formulario al reporte. También se agregaron las fórmulas
