+++ app/src/main/res/layout/activity_actividad_reporte.xml
@@ -6,7 +6,18 @@
     android:paddingBottom="@dimen/activity_vertical_margin"
     tools:context="com.example.miguel.semanai.ActividadReporte">
 
-    <TextView android:text="hola" android:layout_width="wrap_content"
+    <TextView
+        android:text="hola"
+        android:id="@+id/firstMessage"
+        android:layout_width="wrap_content"
         android:layout_height="wrap_content" />
+
+    <TextView
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="New Text"
+        android:id="@+id/test2"
+        android:layout_below="@+id/firstMessage"
+        android:layout_alignParentStart="true" />
 
 </RelativeLayout>
